requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
PyJWT
passlib[bcrypt]
python-multipart
pymongo[srv]==3.11